+++ src/plugin-v3.js
@@ -553,7 +553,7 @@ function transformCss(ast, { env }) {
  * @param {TransformerContext} param1
  */
 function transformAstro(ast, { env, changes }) {
-  let { staticAttrs } = env.customizations
+  let { staticAttrs, dynamicAttrs } = env.customizations
 
   if (
     ast.type === 'element' ||
@@ -569,6 +569,13 @@ function transformAstro(ast, { env, changes }) {
         attr.value = sortClasses(attr.value, {
           env,
         })
+      } else if (
+        dynamicAttrs.has(attr.name) &&
+        attr.type === 'attribute' &&
+        attr.kind === 'expression' &&
+        typeof attr.value === 'string'
+      ) {
+        transformDynamicJsAttribute(attr, env)
       }
     }
   }
@@ -896,6 +903,7 @@ export const parsers = {
     ? {
         astro: createParser('astro', transformAstro, {
           staticAttrs: ['class'],
+          dynamicAttrs: ['class:list'],
         }),
       }
     : {}),